working on redux part, trying to add authorization across different components
